Convert terminal from bottom panel to a tab
